+++ app/Http/Controllers/Library/LendingController.php
@@ -13,6 +13,7 @@ use App\Http\Requests\Library\StoreLendBookToPerson;
 use App\Http\Requests\Library\StoreReturnBookFromPerson;
 use App\Http\Requests\Library\StoreExtendBook;
 use App\Http\Requests\Library\StoreExtendBookToPerson;
+use App\Http\Requests\StorePerson;
 
 class LendingController extends Controller
 {
@@ -35,6 +36,22 @@ class LendingController extends Controller
                 $query->whereNull('returned_date');
             })->get()->sortBy('fullName'),
         ]);
+    }
+
+    public function storePerson(StorePErson $request) {
+        $this->authorize('create', Person::class);
+
+        $person = new Person();
+		$person->name = $request->name;
+        $person->family_name = $request->family_name;
+        $person->gender = $request->gender;
+		$person->date_of_birth = $request->date_of_birth;
+		$person->police_no = !empty($request->police_no) ? $request->police_no : null;
+        $person->nationality = !empty($request->nationality) ? $request->nationality : null;
+        $person->save();
+        
+		return redirect()->route('library.lending.person', $person)
+				->with('success', __('people.person_registered'));		
     }
 
     public function person(Person $person) {