+++ app/Policies/PersonPolicy.php
@@ -60,7 +60,8 @@ class PersonPolicy
      */
     public function create(User $user)
     {
-        return $user->hasPermission('people.manage');
+        return $user->hasPermission('people.manage') 
+            || $user->hasPermission('library.operate');
     }
 
     /**