+++ resources/views/components/form/bsAutocompleteWithButton.blade.php
@@ -1,0 +1,6 @@
+<div class="input-group mb-3">
+    @include('components.form.include.autocomplete')
+    <div class="input-group-append">
+        <button class="btn btn-outline-secondary" type="button" @isset($button_id)id="{{ $button_id }}"@endisset>@isset($button_icon)@icon({{ $button_icon }})<span class="d-none d-sm-inline"> @endisset{{ $button_label ?? 'Button' }}@isset($button_icon)</span>@endisset</button>
+    </div>
+</div>